fix: merge all code in one file for step free
