--- client/src/components/SignInSteps/JobCategory.css
@@ -1,49 +1,0 @@
-.job-category-input {
-  appearance: none;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 0.5rem;
-
-  width: 19.3125rem;
-  height: 5.3125rem;
-  background-color: var(--card-background);
-  border: solid 1px var(--text-color);
-  border-radius: 15px;
-  box-shadow: 0px 4px 4px var(--box-shadow);
-}
-.job-category-box {
-  position: relative;
-  border-radius: 15px;
-  z-index: 2;
-}
-.data-jobs {
-  position: absolute;
-  top: 4px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 0.5rem;
-  z-index: 1;
-}
-.data-jobs h3 {
-  font-size: var(--text-font-size);
-  margin: 15px 0 0 25px;
-}
-.data-jobs span {
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  margin: 0 0px 15px 25px;
-}
-
-.job-category-input:hover,
-.job-category-box:hover {
-  cursor: pointer;
-  box-shadow: none;
-  background-color: var(--card-selected-background);
-}
-
-.active {
-  background-color: var(--card-selected-background) !important;
-}
