+++ client/src/components/SignInSteps/StepThree.css
@@ -1,0 +1,139 @@
+.step-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  background: var(--background);
+  padding-top: 5.68rem;
+  gap: 4rem;
+}
+.step-container h2 {
+  font-size: var(--title-font-size);
+}
+
+.form-job-category {
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+}
+
+.list-job-category {
+  display: flex;
+  flex-direction: column;
+  gap: 0.8rem;
+}
+
+.search-bar {
+  display: flex;
+  flex-direction: column;
+  position: relative;
+}
+
+.search-bar input {
+  height: 3.1875rem;
+  border: 1px solid var(--description-color);
+  background-color: var(--background);
+  font-size: var(--button-font-size);
+  color: var(--description-color);
+  border-radius: 25.5px;
+  padding-left: 25px;
+}
+.error-message {
+  color: red;
+  position: absolute;
+  z-index: 5;
+  bottom: 255px;
+  right: 60px;
+}
+.search-bar label {
+  position: absolute;
+  left: 25px;
+  bottom: 43px;
+  background-color: var(--background);
+  padding: 0 2px;
+}
+.job-category-box {
+  position: relative;
+  border-radius: 15px;
+  z-index: 2;
+}
+.job-category-input {
+  appearance: none;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  gap: 0.5rem;
+
+  width: 19.3125rem;
+  height: 5.3125rem;
+  background-color: var(--card-background);
+  border: solid 1px var(--text-color);
+  border-radius: 15px;
+  box-shadow: 0px 4px 4px var(--box-shadow);
+}
+
+.data-jobs {
+  position: absolute;
+  top: 4px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  gap: 0.5rem;
+  z-index: 1;
+}
+.data-jobs h3 {
+  font-size: var(--text-font-size);
+  margin: 15px 0 0 25px;
+}
+.data-jobs span {
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  margin: 0 0px 15px 25px;
+}
+
+.job-category-input:hover,
+.job-category-box:hover {
+  cursor: pointer;
+  box-shadow: none;
+  background-color: var(--card-selected-background);
+}
+
+.active {
+  background-color: var(--card-selected-background) !important;
+}
+
+.steps-form {
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+}
+.footer-form {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  justify-content: center;
+  align-items: center;
+}
+.footer-form span {
+  color: var(--text-color);
+}
+
+.footer-form ul {
+  list-style-type: none;
+  display: flex;
+  text-align: center;
+  justify-content: center;
+  align-items: center;
+  gap: 1rem;
+}
+
+.footer-form li {
+  width: 8px;
+  height: 8px;
+  border-radius: 50%;
+  background-color: var(--pagination-color);
+}
+.current-step {
+  background-color: var(--description-color) !important;
+}
