--- client/src/components/SignInSteps/StepThree/Three.css
@@ -1,35 +1,0 @@
-.form-job-category {
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-}
-
-.list-job-category {
-  display: flex;
-  flex-direction: column;
-  gap: 0.8rem;
-}
-
-.search-bar {
-  display: flex;
-  flex-direction: column;
-  position: relative;
-}
-
-.search-bar input {
-  height: 3.1875rem;
-  border: 1px solid var(--description-color);
-  background-color: var(--background);
-  font-size: var(--button-font-size);
-  color: var(--description-color);
-  border-radius: 25.5px;
-  padding-left: 25px;
-}
-
-.search-bar label {
-  position: absolute;
-  left: 25px;
-  bottom: 43px;
-  background-color: var(--background);
-  padding: 0 2px;
-}
